// backend/src/modules/integrations/banking/domain/enums/banking.enums.ts
export enum BankProvider {
  ITAU = "ITAU",
  BRADESCO = "BRADESCO",
  SANTANDER = "SANTANDER",
  BANCO_DO_BRASIL = "BANCO_DO_BRASIL",
  CAIXA = "CAIXA",
  SICREDI = "SICREDI",
  SICOOB = "SICOOB",
  INTER = "INTER",
  C6 = "C6",
  NUBANK_PJ = "NUBANK_PJ",
}

export enum PaymentMethod {
  PIX = "PIX",
  BOLETO = "BOLETO",
  CNAB = "CNAB",
  OPEN_FINANCE = "OPEN_FINANCE",
}